requirements.txt
altair
pandas
streamlit
db-dtypes
google-cloud-bigquery
google-cloud-storage
plotly.express
streamlit_option_menu
